Move subtensor.py to bittensor.core and fix related test
