+++ tests/unit_tests/extrinsics/test_network.py
@@ -24,7 +24,7 @@ from bittensor.extrinsics.network import (
     set_hyperparameter_extrinsic,
     register_subnetwork_extrinsic,
 )
-from bittensor.subtensor import Subtensor
+from bittensor.core.subtensor import Subtensor
 
 
 # Mock the bittensor and related modules to avoid real network calls and wallet operations
